apply orthogonal init to all
