+++ jorldy/core/network/rainbow.py
@@ -1,9 +1,8 @@
 import torch
 import torch.nn.functional as F
-import numpy as np
 
 from .base import BaseNetwork
-from .noisy import noisy_l, init_weights
+from .utils import noisy_l, init_weights, orthogonal_init
 
 
 class Rainbow(BaseNetwork):
@@ -30,6 +29,8 @@ class Rainbow(BaseNetwork):
         self.mu_w_v2, self.sig_w_v2, self.mu_b_v2, self.sig_b_v2 = init_weights(
             (D_hidden, N_atom), noise_type
         )
+        
+        orthogonal_init(self.l)
 
     def forward(self, x, is_train):
         x = super(Rainbow, self).forward(x)
